+++ src/sys/hw/mcu/avr/_.rs
@@ -5,6 +5,7 @@
 
 crate::mods_in! {
     mod_ atmega328p;
+    mod pin;
     mod port;
     mod register;
     // mod timer;
@@ -14,6 +15,7 @@ crate::mods_out! { // _mods
     _mods {
         pub use super::{
             atmega328p::_all::Atmega328p,
+            pin::AvrPin,
             port::AvrPort,
             register::AvrReg8,
             // timer::AvrTimer,